+++ changelog.md
@@ -4,6 +4,7 @@ Upcoming (TBD)
 Breaking Changes
 * Change default table format to `mysql_unicode`.
 * Remove support for deprecated environment variable `$DSN`.
+* Remove support for deprecated environment variable `$MYSQL_UNIX_PORT`.
 
 
 Internal